backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.10.6
python-dotenv==1.0.1
loguru==0.7.2
langchain==0.3.19
langchain-openai==0.3.7
langchain-groq==0.2.4
langgraph==0.3.2
sse-starlette==1.6.5
aiohttp==3.9.3 